refactor: move label_filter into src/labels/ subpackage
src/label_filter.py → src/labels/filter.py
tests/test_label_filter.py → tests/labels/test_filter.py

Updates all imports in src/babel_utils.py, src/node.py, and the test
file; updates path references in CLAUDE.md.
